Fix rebasing issues
