+++ crates/trigger/Cargo.toml
@@ -26,7 +26,6 @@ spin-factor-sqlite = { path = "../factor-sqlite" }
 spin-factor-wasi = { path = "../factor-wasi" }
 spin-factors = { path = "../factors" }
 spin-factors-executor = { path = "../factors-executor" }
-spin-runtime-config = { path = "../runtime-config" }
 spin-telemetry = { path = "../telemetry" }
 tokio = { version = "1.23", features = ["fs", "rt"] }
 toml = "0.8"
